Fix: Add a simple check on the passed nodes
When a user passes a node name that doesn't exist but is a substring of
an existing node name, then the script will partially remove the node,
not terminating it but still running Ansible on it. We end up with a
broken setup. The change adds a check at the beginning which makes sure
all host names passed in the --nodes option actually exist as instances
in the cluster network. If not, we exit.