Disable the s2n mlock feature
On OL8 and with slurm 25.11.0 specifically a missing RLIMIT for mlock
(or maybe a too low setting thereof) leads to the s2n plugin not
initializing and slurm crashing (due to a bug in slurm's plugin code).
We don't really need to use the mlock() feature of s2n on our clusters.
If necessary we can look into better configuration later.